requirements.txt
Django~=5.1.1
django-allauth~=64.2.1